refactor(runtime): extract orchestrator tests to separate file
Move the 1,787-line #[cfg(test)] module from orchestrator.rs into
orchestrator_tests.rs using a #[path] attribute. This is a pure file
move with no logic changes — all 45 tests continue to pass.

orchestrator.rs shrinks from 4,701 to 2,916 lines, making it easier
to navigate the production code.
